n3ml f darkmode todo

+++ GrindTracker/varstodo.php
@@ -1,5 +1,4 @@
 <?php 
-
 session_start();
 require_once('connection.php');
 $id = $_SESSION["id"];
@@ -7,8 +6,8 @@ $task = $_POST['task'];
 $timecalendar = $_POST['timecalendar'];
 
 
+
 if (!isset($_POST['delete'])){
-    //add
     $sql="INSERT INTO pr$id (TODO,PrDate,Completed) values('$task','$timecalendar',FALSE)";
 
     $result = mysqli_query($conn, $sql);
@@ -19,7 +18,7 @@ if (!isset($_POST['delete'])){
         echo 0;
     }}
 
-else{
+else if ($_POST['delete'] == "TRUE"){
     //delete
     $sql="DELETE FROM pr$id WHERE TODO='$task' AND PrDate='$timecalendar'";
     $result = mysqli_query($conn, $sql);
@@ -28,12 +27,18 @@ else{
         echo 1;
     } else {
         echo 0;
+    }}
+
+else if ($_POST['delete'] == "FALSE"){
+    //complete
+    $sql="UPDATE pr$id SET Completed = TRUE WHERE TODO='$task' AND PrDate='$timecalendar'";
+    $result = mysqli_query($conn, $sql);
+    if ($result) {
+        echo 1;
+    } else {
+        echo 0;
+    }
 }
-
-}
-
-
-
 
 ?>
